refactor(rules): Remove 'java-001' rule and references in Dockerfile & README :recycle:
A particular security rule, 'java-001' relating to a missing content trust, has been deleted from Dockerfile. In addition, all references to this rule were also removed from the README file. This might be due to redundancy or because the rule is no longer relevant, with 'java-001' no longer available for application to Dockerfiles.