feat(rules): Update and expand Dockerfile security rules :sparkles:
This commit revises the description of an existing security rule, pointing out the recommended use of a non-root user. Additionally, more security rules are added to address other potential issues, such as the use of deprecated sentences, insecure run options, handling of ARG and HEALTHCHECK instructions, and use of the 'latest' tag.

Closes: #2